Renamed `CSimulation` to `CSolver`.
