+++ gillespy2/solvers/cpp/ssa_c_solver.py
@@ -4,9 +4,9 @@ from gillespy2.solvers.cpp.c_decoder import BasicSimDecoder
 from gillespy2.solvers.utilities import solverutils as cutils
 from gillespy2.core import GillesPySolver, gillespyError, Model
 
-from .c_simulation import CSimulation, SimulationReturnCode
+from .c_solver import CSolver, SimulationReturnCode
 
-class SSACSolver(GillesPySolver, CSimulation):
+class SSACSolver(GillesPySolver, CSolver):
     name = "SSACSolver"
     type = "SSASimulation"
