+++ gillespy2/solvers/cpp/tau_leaping_c_solver.py
@@ -1,12 +1,10 @@
-import numpy
-
 from gillespy2.solvers.cpp.c_decoder import BasicSimDecoder
 from gillespy2.solvers.utilities import solverutils as cutils
 from gillespy2.core import GillesPySolver, gillespyError, Model
 
-from .c_simulation import CSimulation, SimulationReturnCode
+from .c_solver import CSolver, SimulationReturnCode
 
-class TauLeapingCSolver(GillesPySolver, CSimulation):
+class TauLeapingCSolver(GillesPySolver, CSolver):
     name = "TauLeapingCSolver"
     type = "TauLeapingSimulation"
